Add files via upload
A URL shortener is a service that converts a long URL (Uniform Resource Locator) into a shortened form, which is easier to share, type, and remember. The shortened URL is redirected to the original, longer URL when it is clicked, making it possible to access the original web page through the short link.

URL shorteners are often used to save space when sharing links on social media platforms, where there is a character limit for posts, or to simplify the appearance of a long, complex URL. They are also used to track clicks on links, allowing the user to see how many clicks a particular link has received and where they are coming from.